Chatbot/requirements.txt
flask==2.3.3
google-generativeai==0.3.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3 